One float made the .rpm uninstallable; that, and three lies in the packaging
Reported: `dnf install` refusing v0.10.0 with

    nothing provides libm.so.6(GLIBC_2.43)(64bit) needed by cordial-0.10.0

**One line, in our own code.** `native/opensles.cpp` computed a millibel
volume with `std::log10` on a `float`. glibc 2.43 added a correctly-rounded
`log10f` and made it the default binding, so building on a 2.43 host emitted
`log10f@GLIBC_2.43` -- the only symbol in either binary above 2.39. The result
is cast to an integer millibel two lines later, so nothing there wanted single
precision; `log10` has been GLIBC_2.2.5 since before any of this. Both binaries
now top out at 2.39, measured with `objdump -T`.

The AppImage had it too, and that one bundles no libc, so the format whose
whole purpose is running anywhere would have run on Fedora 44 and nothing else.
Nobody noticed because every builder and every developer machine is newer than
every user.

`packaging/check-glibc-floor.sh` refuses a binary above the floor and names the
symbols responsible. Wired into the AppImage and .deb jobs. Its control: with
the floor set to 2.38 it fails and prints `pidfd_getpid (GLIBC_2.39)`. A
container's glibc moving forward changes no line in this repository, which is
exactly why this wants a check and not a habit.

**Three things the packages told users that were false**, the first reported
today as "arch message for cordial has no apk and cant fetch one is false now":

  - "CORDIAL SHIPS NO ROBLOX BUILD AND CANNOT DOWNLOAD ONE" -- it has been able
    to download and signature-check one since v0.9.0. In the Arch .install, the
    deb postinst, the deb control, and the rpm description.
  - "text fields do not draw what you type" -- solved and verified on
    2026-08-25.
  - "the pointer is not captured in first person" -- pointer capture exists.

All four now say what is true, including what is still broken: voice chat, and
the signed-in startup freeze. A package description that undersells is not
safer than one that oversells, it is just wrong in the other direction.

**`cordial` is now the command**, asked for today: nobody wants to type the
second word and every other launcher answers to its own name. A symlink beside
`cordial-shell` in all five formats rather than a rename, so anything already
invoking the old name keeps working. `cordial-run` deliberately gets no alias.

One self-inflicted bug caught before it shipped: `%{_bindir}/cordial` first
landed in the spec's `%install` rather than `%files`, which would have run
`/usr/bin/cordial` as a command during the build. That is the same shape as the
`install -Dpm 0644 NOTICE NOTICE` self-copy this file had before.

+++ native/opensles.cpp
@@ -809,8 +809,21 @@ SLresult volume_GetVolumeLevel(SLVolumeItf self, SLmillibel* pLevel) {
     if (linear <= 0.0f) {
         *pLevel = SL_MILLIBEL_MIN;
     } else {
-        float mb = 2000.0f * std::log10(linear);
-        *pLevel = static_cast<SLmillibel>(std::max<float>(SL_MILLIBEL_MIN, std::min<float>(SL_MILLIBEL_MAX, mb)));
+        // **`double`, and the reason is packaging rather than precision.**
+        // glibc 2.43 added a correctly-rounded `log10f` and made it the
+        // default binding, so building this line against a 2.43 host emitted
+        // `log10f@GLIBC_2.43` -- and that one symbol, the only one in either
+        // binary above 2.39, made the `.rpm` refuse to install anywhere older
+        // with "nothing provides libm.so.6(GLIBC_2.43)(64bit)". It would have
+        // done the same to the AppImage, which bundles no libc and is supposed
+        // to be the portable one.
+        //
+        // `log10` has been `GLIBC_2.2.5` since before any of this and there is
+        // no version of glibc without it. The result is cast to an integer
+        // millibel two lines down, so nothing here wanted single precision in
+        // the first place.
+        double mb = 2000.0 * std::log10(static_cast<double>(linear));
+        *pLevel = static_cast<SLmillibel>(std::max<double>(SL_MILLIBEL_MIN, std::min<double>(SL_MILLIBEL_MAX, mb)));
     }
     return SL_RESULT_SUCCESS;
 }
